refactor: follow naming convention for naming files
