+++ app/(dashboard)/mailboxes/[mailbox_slug]/[category]/conversation/messageActions.tsx
@@ -7,9 +7,9 @@ import { DraftedEmail } from "@/app/types/global";
 import { triggerConfetti } from "@/components/confetti";
 import { useFileUpload } from "@/components/fileUploadContext";
 import { GenerateKnowledgeBankDialog } from "@/components/generateKnowledgeBankDialog";
-import { useExpiringLocalStorage } from "@/components/hooks/use-expiring-local-storage";
-import { toast } from "@/components/hooks/use-toast";
+import { useExpiringLocalStorage } from "@/components/hooks/useExpiringLocalStorage";
 import { useSpeechRecognition } from "@/components/hooks/useSpeechRecognition";
+import { toast } from "@/components/hooks/useToast";
 import { KeyboardShortcut } from "@/components/keyboardShortcut";
 import LabeledInput from "@/components/labeledInput";
 import TipTapEditor, { type TipTapEditorRef } from "@/components/tiptap/editor";
