+++ app/(dashboard)/mailboxes/[mailbox_slug]/[category]/ticketCommandBar/mainPage.tsx
@@ -12,7 +12,7 @@ import {
 import { useMemo, useRef } from "react";
 import { useConversationContext } from "@/app/(dashboard)/mailboxes/[mailbox_slug]/[category]/conversation/conversationContext";
 import { Tool } from "@/app/(dashboard)/mailboxes/[mailbox_slug]/[category]/ticketCommandBar/toolForm";
-import { toast } from "@/components/hooks/use-toast";
+import { toast } from "@/components/hooks/useToast";
 import useKeyboardShortcut from "@/components/useKeyboardShortcut";
 import { api } from "@/trpc/react";
 import GitHubSvg from "../icons/github.svg";
