+++ app/(dashboard)/mailboxes/[mailbox_slug]/dashboard/dashboardContent.tsx
@@ -5,7 +5,7 @@ import { useState } from "react";
 import { DateRange } from "react-day-picker";
 import { PeopleTable } from "@/app/(dashboard)/mailboxes/[mailbox_slug]/dashboard/peopleTable";
 import { ReactionsChart } from "@/app/(dashboard)/mailboxes/[mailbox_slug]/dashboard/reactionsChart";
-import { useIsMobile } from "@/components/hooks/use-mobile";
+import { useIsMobile } from "@/components/hooks/useMobile";
 import { PageHeader } from "@/components/pageHeader";
 import { Panel } from "@/components/panel";
 import { DashboardAlerts } from "./dashboardAlerts";
