+++ app/(dashboard)/mailboxes/[mailbox_slug]/settings/customers/autoCloseSetting.tsx
@@ -1,8 +1,8 @@
 "use client";
 
 import { useState } from "react";
-import { toast } from "@/components/hooks/use-toast";
 import { useSavingIndicator } from "@/components/hooks/useSavingIndicator";
+import { toast } from "@/components/hooks/useToast";
 import { SavingIndicator } from "@/components/savingIndicator";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
