+++ app/(dashboard)/mailboxes/[mailbox_slug]/settings/preferences/mailboxNameSetting.tsx
@@ -1,8 +1,8 @@
 "use client";
 
 import { useState } from "react";
-import { toast } from "@/components/hooks/use-toast";
 import { useSavingIndicator } from "@/components/hooks/useSavingIndicator";
+import { toast } from "@/components/hooks/useToast";
 import { SavingIndicator } from "@/components/savingIndicator";
 import { Input } from "@/components/ui/input";
 import { useDebouncedCallback } from "@/components/useDebouncedCallback";
