+++ app/(dashboard)/mailboxes/[mailbox_slug]/settings/tools/toolSetting.tsx
@@ -2,7 +2,7 @@
 
 import { PlusCircle } from "lucide-react";
 import { useEffect, useState } from "react";
-import { toast } from "@/components/hooks/use-toast";
+import { toast } from "@/components/hooks/useToast";
 import { Button } from "@/components/ui/button";
 import { api } from "@/trpc/react";
 import SectionWrapper from "../sectionWrapper";
