+++ components/tiptap/image.tsx
@@ -1,7 +1,7 @@
 import { mergeAttributes, Node as TiptapNode } from "@tiptap/core";
 import { NodeViewWrapper, ReactNodeViewRenderer, type NodeViewProps } from "@tiptap/react";
 import { useCallback, useEffect, useRef, useState } from "react";
-import { toast } from "@/components/hooks/use-toast";
+import { toast } from "@/components/hooks/useToast";
 import { useRunOnce } from "@/components/useRunOnce";
 
 export const imageFileTypes = ["image/jpeg", "image/jpg", "image/png", "image/gif", "image/heic"];
