+++ components/useToolExecution.ts
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import { useConversationContext } from "@/app/(dashboard)/mailboxes/[mailbox_slug]/[category]/conversation/conversationContext";
-import { toast } from "@/components/hooks/use-toast";
+import { toast } from "@/components/hooks/useToast";
 import { api } from "@/trpc/react";
 
 export const useToolExecution = () => {
